Hackathon/RFAPI_DB/requirements.txt
numpy>=1.21.0
opencv-python>=4.8.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.5.0
facenet-pytorch>=2.5.0
flask>=2.0.0
python-socketio>=5.0.0
eventlet>=0.30.0
mysql-connector-python>=9.0.0